+++ src/proxy_protect/components/rules.clj
@@ -1,32 +1,37 @@
 (ns proxy-protect.components.rules
   (:require [clojure.tools.logging :refer [info]]
             [com.stuartsierra.component :as component]
-            [jsonista.core :as json]
-            [org.httpkit.client :as client]))
+            [cheshire.core :as json]))
 
-(defmulti request-type :request-method)
-(defmethod request-type :get [_] client/get)
-(defmethod request-type :post [_] client/post)
-(defmethod request-type :put [_] client/put)
-(defmethod request-type :delete [_] client/delete)
-(defmethod request-type :patch [_] client/patch)
+(defn http-method-allowed?
+  [allowed-methods request-method]
+  (let [string-request-method (name request-method)]
+    (some #(or (= % "any")
+               (= % string-request-method))
+          allowed-methods)))
 
-(defn proxy-request
-  [request destination]
-  (let [request-fn (request-type request)
-        response @(request-fn destination {:throw-exceptions false})]
-    (assoc response :headers nil)))
+(defmulti create-match-fn :type)
+(defmethod create-match-fn "absolute"
+  [rule]
+  (let [allowed-methods (:allowedMethods rule)]
+    (assoc rule :match-fn (fn [request]
+                            (let [{:keys [request-method  uri]} request]
+                              (and (http-method-allowed? allowed-methods
+                                                         request-method)
+                                   (= uri (:source rule))))))))
+(defmethod create-match-fn "regex"
+  [rule]
+  (let [pattern (re-pattern (:source rule))
+        allowed-methods (:allowedMethods rule)]
+    (assoc rule :match-fn (fn [request]
+                            (let [{:keys [request-method  uri]} request]
+                              (and (http-method-allowed? allowed-methods
+                                                         request-method)
+                                   (re-matches pattern uri)))))))
 
 (defn- rules->edn
   [rules]
-  (let [object-mapper (json/object-mapper {:decode-key-fn true})]
-    (json/read-value rules object-mapper)))
-
-(defn- rule->rule-fn
-  [rule]
-  (assoc rule :rule-fn (fn [request]
-                         (let [uri (:uri request)]
-                           (= uri (:source rule))))))
+  (json/decode rules true))
 
 (defrecord Rules [rules]
   component/Lifecycle
@@ -35,7 +40,7 @@
     (assoc component :rules (->> rules
                                  slurp
                                  rules->edn
-                                 (mapv rule->rule-fn))))
+                                 (mapv create-match-fn))))
   (stop [component]
     (info "Stopping the Rules Component...")
     (assoc component :rules nil)))