+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-## [Unreleased] - 2020-09-07
+## [Unreleased] - 2020-09-09
 
 ### Added
 - Rules Component
 - Webserver Component
 - Wrap Proxy Middleware
+- Protect Routes by HTTP Method
+- Regex and Absolute Path Match for Source URI
+
+### Changed
+- Now using log4j2
+- Using `clj-http` over `http-kit`